% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ALL_down_exp_TSGs}
\alias{ALL_down_exp_TSGs}
\title{Download the down-regulated tumor suppressor genes in tumor samples versus
normal tissue}
\format{
A list of tumor suppressor:
\describe{
  \item{PANCAN}{All the down regulated 983 human TSGs are lower expressed in
  cancer samples of TCGA comparing to the TCGA normal tissue samples,
  GeneID: Entrz Gene ID,
  GeneName: HGNC Gene Symbol,
  NumberOfCancers: Number Of Cancer type,
  CancerTypes: Cancer Type.
  }
  \item{BLCA, BRCA, COAD, HNSC, KICH, KIRC, LUAD, LUSC, PRAD, THCA, UCEC}{
   GeneID: Entrz Gene ID,
   GeneSymbol: HGNC Gene Symbol}
  }
}
\source{
https://bioinfo.uth.edu/TSGene/index.html
}
\usage{
data(ALL_down_exp_TSGs)
}
\description{
Download the down-regulated tumor suppressor genes in tumor samples versus
normal tissue
}
\keyword{datasets}
