% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanTSGs}
\alias{HumanTSGs}
\title{All the 1217 human tumor suppressor genes with basic annotations}
\format{
A tibble data frame:
\describe{
  \item{GeneID}{Entrz Gene ID}
  \item{GeneSymbol}{HGNC Gene Symbol}
  \item{Alias}{Aliases}
  \item{Links}{External Website}
  \item{Chromosome}{Chromosome}
  \item{Cytoband}{Chromosome postion}
  \item{Description}{Gene functional annotation}
  \item{GeneType}{Gene biological Type}
  }
}
\source{
https://bioinfo.uth.edu/TSGene/index.html
}
\usage{
data(HumanTSGs)
}
\description{
All the 1217 human tumor suppressor genes with basic annotations
}
\keyword{datasets}
