% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig_exp}
\alias{sig_exp}
\title{All the precomputed P-values for gene expression difference between tumor and
normal samples for all human tumor suppressor genes in pan-cancer}
\format{
A tibble data frame:
\describe{
  \item{GeneID}{Entrz Gene ID}
  \item{Symbol}{HGNC Gene Symbol}
  \item{BLCA}{Bladder urothelial carcinoma}
  \item{BRCA}{Breast invasive carcinoma}
  \item{COAD}{Colon adenocarcinoma}
  \item{HNSC}{Head and Neck squamous cell carcinoma}
  \item{KICH}{Kidney chromophobe}
  \item{KIRC}{Kidney renal clear cell carcinoma}
  \item{LUAD}{Lung adenocarcinoma}
  \item{LUSC}{Lung squamous cell carcinoma}
  \item{PRAD}{Prostate adenocarcinoma}
  \item{THCA}{Thyroid carcinoma}
  \item{UCEC}{Uterine corpus endometrial carcinoma}
  }
}
\source{
https://bioinfo.uth.edu/TSGene/index.html
}
\usage{
data(sig_exp)
}
\description{
All the precomputed P-values for gene expression difference between tumor and
normal samples for all human tumor suppressor genes in pan-cancer
}
\keyword{datasets}
